Primary Text & Button Styled + comments
The primary message text has been styled, as well as
the button, to better reflect what's required in the reference.
Previously, there was classes which were not provided comments
throughout the CSS file. This has been rectified.

=== Class Specific Changes ===
:root
The variable bgDropShadow was created for both the credentials
    and log-in button.

span
The font was italicized to have two words reflect the reference.

.msg_action_title, .msg_bork
Styled where both will have the same size and weight.

.msg_action_title
Removed the margin-top so that there's no longer any
    extra space due to the padding given to its parent container.

.credentials
The padding property had its lost value dropped since the left
    padding is being provided by .container_sign_up > div.

    The box-shadow property was updated to have the same value
    as the .log-in > button.

.log-in > button
Created to get the button styled to be as close to the reference
    image as possible.